backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
asyncpg
pydantic
pydantic-settings
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
httpx
jinja2
openapi-generator-cli
pyyaml
jsonschema
pytest
pytest-asyncio
pytest-cov
factory-boy
redis
celery
flower
prometheus-client
structlog
psutil
